interp: Move _registry to decode
More related to decode than interp. Maybe should have a format.go?
