+++ pkg/interp/interp.go
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"io"
 	"io/fs"
-	"io/ioutil"
 	"math/big"
 	"path/filepath"
 	"strconv"
@@ -61,7 +60,6 @@ func init() {
 			{"_extkeys", 0, 0, i._extKeys, nil},
 			{"_exttype", 0, 0, i._extType, nil},
 			{"_global_state", 0, 1, i.makeStateFn(i.state), nil},
-			{"_registry", 0, 0, i._registry, nil},
 			{"history", 0, 0, i.history, nil},
 			{"_display", 1, 1, nil, i._display},
 			{"_can_display", 0, 0, i._canDisplay, nil},
@@ -536,86 +534,6 @@ func (i *Interp) makeStateFn(state *interface{}) func(c interface{}, a []interfa
 			*state = a[0]
 		}
 		return *state
-	}
-}
-
-func (i *Interp) _registry(c interface{}, a []interface{}) interface{} {
-	uniqueFormats := map[string]decode.Format{}
-
-	groups := map[string]interface{}{}
-	formats := map[string]interface{}{}
-
-	for fsName := range i.registry.Groups {
-		var group []interface{}
-
-		for _, f := range i.registry.MustGroup(fsName) {
-			group = append(group, f.Name)
-			if _, ok := uniqueFormats[f.Name]; ok {
-				continue
-			}
-			uniqueFormats[f.Name] = f
-		}
-
-		groups[fsName] = group
-	}
-
-	for _, f := range uniqueFormats {
-		vf := map[string]interface{}{
-			"name":        f.Name,
-			"description": f.Description,
-			"probe_order": f.ProbeOrder,
-			"root_name":   f.RootName,
-			"root_array":  f.RootArray,
-		}
-
-		var dependenciesVs []interface{}
-		for _, d := range f.Dependencies {
-			var dNamesVs []interface{}
-			for _, n := range d.Names {
-				dNamesVs = append(dNamesVs, n)
-			}
-			dependenciesVs = append(dependenciesVs, dNamesVs)
-		}
-		if len(dependenciesVs) > 0 {
-			vf["dependencies"] = dependenciesVs
-		}
-		var groupsVs []interface{}
-		for _, n := range f.Groups {
-			groupsVs = append(groupsVs, n)
-		}
-		if len(groupsVs) > 0 {
-			vf["groups"] = groupsVs
-		}
-
-		if f.Files != nil {
-			files := map[string]interface{}{}
-
-			entries, err := f.Files.ReadDir(".")
-			if err != nil {
-				return err
-			}
-
-			for _, e := range entries {
-				f, err := f.Files.Open(e.Name())
-				if err != nil {
-					return err
-				}
-				b, err := ioutil.ReadAll(f)
-				if err != nil {
-					return err
-				}
-				files[e.Name()] = string(b)
-			}
-
-			vf["files"] = files
-		}
-
-		formats[f.Name] = vf
-	}
-
-	return map[string]interface{}{
-		"groups":  groups,
-		"formats": formats,
 	}
 }
